requirements.txt
asgiref==3.5.2
attrs==22.1.0
dj-database-url==1.0.0
Django==4.1
django-filter==22.1
django-herokuify==1.0rc3
django-on-heroku==1.1.2
django-pytest==0.2.0
Faker==14.1.0
flake8==5.0.4
gunicorn==20.1.0
iniconfig==1.1.1
mccabe==0.7.0
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.9.1
pyflakes==2.5.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
whitenoise==6.2.0
